Extract App component back into its own file
