+++ src/options/index.test.js
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom/extend-expect";
-import App from ".";
+import App from "./App";
 import React from "react";
 import { render } from "@testing-library/react";
